Add initial -lzs- decompressor.
Unfortunately I've been unable to find a version of LArc that
generates -lzs- data. I'm therefore relying on test data generated
by a script that converts from -lz5- to -lzs-. The decompressor
code appears to work with this test data; however, the data itself
needs to be confirmed to be valid -lzs-. The next stage is to put
the generated data into a .lzs archive and check that DOS versions
of LArc can decompress it successfully.