Correctly handle Amiga self-extracting archives.
Amiga LhASFX v1.2 archives include an embedded LHA file that contains
a description of the command line parameters for the self-extractor.
Because of this, the embedded file would be detected and decoded instead
of the actual file. Add a signature check to detect these archives and
skip over the embedded file.