lz5: Fill initial history buffer correctly.
The -lz5- history buffer contains certain pre-filled patterns which are
useful to copy from. There are archives in the wild (such as lzsfx33.lzs)
that rely on these initial patterns and will not extract correctly without
them, so fill the buffer correctly and add a constructed test archive
that performs copies from the entire history buffer to check that it is
filled correctly.

This fixes #13. Thanks to roytam1 for the bug report.