Replace sed usage with a dedicated program.
commit 0207d320530d467 was a previous attempt at solving this problem
but appears to have been unsuccessful since there are still issues with
the tests on BSD systems. Since the problem originates with sed, a
straightforward solution is to just replace it with a filter program
that does a simple string replace. This appears to fix the problem as
the tests now pass on Linux, macOS and OpenBSD, at least.