Add initial recognition of LHARK archives.
This is a first step towards feature request #20 and adding support for
LHARK archives. They are generated with a space character (' ') as the
OS type, which no other tool appears to do, so we can use this to
identify them. Since we can identify them, we can also use this to
distinguish between normal -lh7- compression and LHARK's version, which
is named the same but a different algorithm. We internally rename this
to -lk7- so that the decompression code can continue to pick the decoder
based on the name.