Add decoder for LHARK's -lh7- format.
Part of #20. We call this format -lk7- to distinguish this from the
usual -lh7-. There are some code hacks needed to support LHARK's
custom tweaks to the compression scheme but otherwise it's largely
the same as -lh5- et. al.

I'm indebted to Jason Summers, his excellent tool DEARK and his
comprehensive writeup on the LHARK compression format found here:
<https://entropymine.wordpress.com/2020/12/24/notes-on-lhark-compression-format/>