Fix Amiga -lh0- directories bug.
The Amiga version of LhA (one of them?) can generate malformed archives
where directories are represented with the -lh0- compression type
instead of -lhd-. Headers of this type are considered invalid and
extraction will stop once one of them is reached. Since this bug is
limited to just Amiga archives, we can construct a targeted workaround
that internally fixes the compression method to the -lhd- type.

Thanks go to @zzarko for originally reporting the bug, @polluks for
supplying an additional example, @alfredone for supplying a very useful
minimal test case archive (included here), and @jsummers for correctly
deducing the underlying bug.

This fixes #31.