Change symlink representation to match Unix LHa.
The way symlinks are represented in archives generated by the original
Unix LHa is very weird and arguably broken. However, we don't want to
invent yet another way of representing them in headers, and if we do
fix the bug, older versions of Lhasa won't be able to read those
archives. This adds some extra complication but I think it's the right
thing to do regardless.