Fix undefined behavior with BitStreamReader.
When we read less than four bytes, the remaining bytes in buf are zero thanks
to the memset() call. We bitwise-or in all the bytes regardless, but the
remainder bytes get a nonsensical shift value because reader->bits > 16. Since
these byte are always zero anyway it doesn't affect the result, but it is still
undefined behavior regardless and ought to be fixed.

Caught by Clang UBSAN and reported by @sezero - thanks! Fixes #49.