Tweak list output to make parsing easier
Until now we've been very careful to ensure our list output was
identical to the output from the original lha tool, since a design goal
for Lhasa was to be a drop-in replacement for that tool. However,
numerous GUI programs parse the output from lha and make incorrect
assumptions about the format. I recently fixed one bug in xarchiver
(ib/xarchiver#214) but have subsequently uncovered similar bugs in
file-roller and other programs derived from it.

Most of these programs assume that the list output is consistent,
and specifically that the UID/GID field always has a value. If the
archive was generated on a non-Unix system then this is not the case;
Unix lha shows whitespace in that column. We can therefore make the
output more consistent by printing a spacer word when UID/GID are
not present. With this change, xarchiver (without my fix mentioned
above) now parses the output correctly; file-roller also seems to open
some archives correctly, although there appears to be an unrelated bug
when opening MS-DOS archives.

I've also applied a similar tweak to the timestamp column, since we
were printing whitespace there too when no timestamp was present (Unix
lha actually prints a timestamp, just an invalid one.

The choice of asterisks ("******") as the spacer word is inspired by
the fact that we already print the same thing in the compression ratio
column when listing directories.

This is part of #59.