Factor out code to fill/read output buffer
Previously this was all just part of `lha_encoder_read()`. We add a new
function, `lha_encoder_fill()`, that just fills up the output buffer to
a target amount without reading any data.

Eventually we'll use this from within the writer code to automatically
select the encoding for a particular file based on its observed
compression ratio (ie.  pre-compressed data should be stored as
`-lh0-`). Anticipating this, this change also adds "headroom" space
within the output buffer so that we will always be able to fill up to a
particular amount.