Compress using `-lh1-`, falling back to `-lh0-`
We can use the new `lha_encoder_fill` function to test the compression
ratio of data at the start of the file being compressed. If we don't
manage to reduce the size by at least 2%, the file is probably already
compressed. In this case, switch to `-lh0-` and store the file
uncompressed.