Only write terminal bits when we are really at EOF
This is a correction to my earlier commit 855e73c0e798cc2cde4eaa68a;
we should only ever write the terminal zero bits when we have definitely
reached the EOF. This fixes a bug with some files where we were
mistakenly inserting the zero bits in the middle of the stream, which
obviously was leading to corruption.

To do this properly we need to tweak the code that refills the input
buffer, to invoke the read callback repeatedly. We only set the EOF flag
when the read callback returns zero bytes, so in the case of a short
read we might otherwise not get that zero result until the next
iteration through the loop.