Bump up size of BitStreamWriter buffer to 64 bits
While BitStreamReader has a callback that allows it to refill itself,
BitStreamWriter only has a relatively small amount of storage that needs
to be periodically flushed. 32 bits is enough for most commands but
during testing I found that some files would trip the limit and overrun
the buffer. 64 bits should be enough for our needs.