Add search buffer checks for most recent bytes
The contents of the search buffer are indexed with hash chains but the
most recent three bytes written to the buffer cannot be indexed because
the hash function requires at least three bytes. So explicitly check the
most recent bytes before we do the hash chain lookups, since we might
get a better result than anything we find in the hash chains.