Add `LHAFileHeader` 64-bit size fields
This adds the new fields but preserves the old fields to retain ABI
compatibity on systems where `size_t` is a 32-bit integer. New code will
use the new fields but old code will keep using the old fields. We
populate both when parsing the file headers.

The main other change here is that the list code needs to use 64-bit
counts for its statistics. This is something we ought to have been doing
anyway, since it was already possible for these to overflow if there was
a .lzh file containing lots of compressed large files.

Part of #60.