Decode file sizes extended header (0x42)
This decodes the 64-bit file sizes extended header that is generated by
the MorphOS port of lha when storing huge files that exceed the 32-bit
limit. Possibly other tools generate archives with this header as well,
since it's listed in the jLHA notes on the LHA format; however, at
present I'm not aware of any.

We keep the ABI compatibility fields in sync but prevent overflow; if
the value cannot be represented, we use `SIZE_T_MAX` instead. At least
any user of these fields will read as many bytes as they can.

To confirm this works, we now add the `h2_huge.lzh` file to the tests.
This contains a 4.7+GiB file just containing a long run of zero bytes.
Since this requires quite a lot of disk space to extract, I haven't
added it to the extract tests.

This fixes #60.