Change `lha_decoder_get_length` to return 64-bit
To keep ABI compatibility we can't really change this, but we can define
a new function that returns a 64-bit value. With some #define magic we
can invisibly redirect any callers over to use our new function instead.
We keep a compatible version of the function under the old name that
continues returning `size_t` for compatibility (on some systems, this is
a 32-bit value).

Part of #60.