Scale up progress block size to prevent overflow
The `LHADecoderProgressCallback` type uses 32-bit values, which in
theory could overflow now that we support 64-bit file sizes. For the
majority of cases we keep using the same block sizes for reporting, but
if the file is really big (100s of MB), scale up the block size.

This ensures that the number of blocks we report has a maximum of 128K,
which ought to be enough detail for anyone anyway (the interface is
intended for reporting progress to humans). This way, the 32-bit
callback parameters can never overflow.

Block sizes have always been opaque to the user, but this also updates
the doc comment to make clear that they're arbitrary and no assumptions
should be made about the size of a block.

A minor follow-up to #60.