Set libtool versioning information for liblhasa
Until now we have always been building the library with the same
filename of `liblhasa.so.0.0.0`, which is the default if you do not
supply any version information. Since we now have at least some users
compiling against the library, we ought to be treating the versioning at
least a little bit more seriously.

The versioning information is supplied on build via the `-version-info`
command line option. The actual values to supply are a bit more arcane
and are not the same as the package version. The guide by flameeyes here
gives some useful rules of thumb:

<https://autotools.info/libtool/version.html>

> Always increase the revision value. Increase the current value
> whenever an interface has been added, removed or changed. Increase the
> age value only if the changes made to the ABI are backward compatible.

Since the v0.4.0 release we have made some ABI changes, but they should
be backwards compatible. Therefore, I've picked 1:1:1 as the initial
values to use for the library version. With this change, we now produce
a library named `liblhasa.so.0.1.1` which looks sensible: the major part
is unchanged (0) while the minor parts have increased.