Revise comments around library/libtool versioning
After reading the libtool documentation I understand a bit better how
this works and what these numbers mean. With this in mind I no longer
think the comments I copy/pasted are the best advice here. Instead,
replace them with a link to the libtool manual: I think the best advice
is to learn how the system is really supposed to work and what the
numbers are supposed to represent.

For the same reason, I've bumped LIBVER_REVISION back down to zero,
following the guidelines from the libtool manual:

> If any interfaces have been added, removed, or changed since the last
> update, increment current, and set revision to 0.