Add 64-bit ABI workaround for `lha_decoder_new`
This is equivalent to the `lha_decoder_get_length` hack put in several
commits back to add 64-bit file support. I missed the fact that the
`lha_decoder_new` function also has a 64-bit parameter for file length,
and because of this, the new test for `h2_huge.lzh` was failing on
systems with 32-bit `size_t`s.

Follow-up to #60.