Make list compression ratio more pessimistic
With the recently added h2_huge.lzh, I noticed that the compressed file
was listed as being 0.0% of the original size, since the compression
ratio is so high. This happens because we were using the regular %f
printf formatting, which presumably rounds to the nearest 0.1%. However,
it nonetheless seems rather misleading.

This instead changes the output to always round upwards to the next
0.1%. This is more pessimistic than the old rounding but seems more
honest. With this change, we never show any file as having a 0.0% ratio,
0.1% is the best that can ever be achieved. A bunch of the expected
outputs from the list command had to be updated to match the new logic.